kern-tools-native: fix fragments merging script
When concatenating kernel config fragments, if a fragment file doesn't
end with a carriage return, then the last config option of the
fragment and the first one of the next fragment are on the same line
in the result file. This leads to misconfiguration because some
options are not taken into account.

This commit includes a patch that will add a \n in the result file
after each fragment.

Bug-AGL: SPEC-1475

Change-Id: I52d13b42909f56f4608860af865a80d0d9b2579e
Signed-off-by: Stephane Desneux <stephane.desneux@iot.bzh>