Fix do_rootfs eats huge time on docker environment
RPM is causing a serious performance regression on Docker.

Docker can set ulimit -n to 1048576 but it causes huge time
consumption when thousands of packages are installed like
bitbake agl-demo-platform.

This issue is already resolved in upstream of RPM and
yocto follows it at sumo, so backporting it into Flounder.

Bug-AGL: SPEC-1622

Change-Id: Ia8d97daea663f9682928a14ab84199ed6fda6d61
Signed-off-by: Tadao Tanikawa <tanikawa.tadao@jp.panasonic.com>