meta-intel: Support the upsquare board
This patch adds the necessary CONFIG for booting the upsquare board.

The upsquare have an R8169 ethernet card and need it built-in for
the kernel DHCP autoconf to work.

The upsquare need some MFD_INTEL configs to boot.

The meta-intel miss some CONFIG necessary for LAVA and NBD to works.

Bug-AGL: SPEC-2075

Change-Id: I990732c78b53ac591dde83f4e372b420ea558fa7
Signed-off-by: Corentin Labbe <clabbe@baylibre.com>