meta-agl-profile-graphical*: Add qtaglextras and dependencies
Add qtaglextras to meta-agl-profile-graphical-qt5, and its
dependencies libhomescreen and agl-service-homescreen to
meta-agl-profile-graphical.  This is required to fix building the
cluster demo SDK now that qtaglextras has been added to
packagegroup-qt5-toolchain-target.  These changes are a partial
backport of the move of the contents of meta-hmi-framework in
master / Halibut.

Bug-AGL: SPEC-2484

Change-Id: I94493c65d2253ceb2204b34a8ad216910851a54b
Signed-off-by: Scott Murray <scott.murray@konsulko.com>