distro-build-manifest: add DIST_BUILD_TS in all manifests
The build timestamp is added to build manifests deployed in target images and SDK.

Also, this patch adds a new option '-t|--timestamp' to distro-manifest-generator.sh
to let the recipe pass the same timestamp as the one used for the image suffix
(in Yocto 2.6/Thud, DATETIME is used). So DIST_BUILD_TS should have the same value
as the image version suffix.

Bug-AGL: SPEC-2649

Change-Id: I2a9d11b2ce59753bca31ebb8e82d46c224110d68
Signed-off-by: Stephane Desneux <stephane.desneux@iot.bzh>