mesa: Upgrade Mesa for Raspberry Pi 4
Bring Mesa 19.1.6 to meta-agl-bsp/meta-raspberrypi because there
are driver issues with Mesa 18.1 for Raspberry Pi 4. The recipes
for this new version are coming from the master branch of layer
openembedded-core. With the newer version of Mesa Weston works
fine on Raspberry Pi 4 and the home screen appears as expected.

Bug-AGL: SPEC-2656

Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>
Change-Id: I0d4f6e6fdff3c6ac54a72d19abfc292efcea7a3f