templates/feature/agl-compositor: Include SRCREVs
It seems that in CI, yocto fails to retrieve the correct
branch, presumably due to the SRCREV pointing to HEAD of the
agl/master branch instead of the one by directly specified.

This patch uses the SRCREV from sandbox/mvlad/agl-compositor to make
it resolves correctly the branch and implicitly the fetch succeeds.

With it, this adds further dependency of wayland-scanner for launcher
application, as to able to activate applications with the
agl-desktop-shell protocol.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Ie6a559be56702843abdb580ec66998b6b4c3f978
(cherry picked from commit 3f746bbb527542ef65e9d9829d4ca00ea1fb9901)