Cleanup of deprecated recipes
We deprecated meta-intel in favour of a combined qemux86-64 image.
Thus we can remove bbappends related to meta-intel.

We can also remove the joule bootdisk (board long EOL).

The intel kernel config fragments are moved to the generic recipes-kernel folder
for a later inspection as to enhance the kernel options available in the
qemux86-64 build.

Finally also remove the override for meta-synopsys (no dunfell branch).

Bug-AGL: SPEC-3468

Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>
Change-Id: I8c999a4f6d3ad4b85c71f94dac13c3563c440a88