recipes-graphics/wayland/agl-compositor: Bump SRCREV for agl-compositor
Bumps the agl-compositor to the latest version. This update removes,
from the default, allow-all policy, the private extensions bind
tests, such that applications making use of those private
extensions are free to bind to them. We'll need some better ways to
corroborate/integrate with AFM to allow a dynamic way to communicate
which applications can bind to those private extensions.

On the other hand, for the deny-all policy, the permitted application
list has been extended to include the cluster dashboard and the cluster
receiver applications, and users can use that in practice if they
require some mechanism in place to control which application can bind to
the private extensions.

In more detail, the changes are:

- c7bc55dce459 layout: Add missing hooks for other surface roles
- bf016d204089 layout: Advertise state change for other types of surface roles
- 35f6733fb39b shell: Assume consistency with the desktop roles
- 5fe16df23483 policy-*: Remove any private binding restrictions for allow-all policy

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I3c25af73a76404c87e09245fee04d0f3a2a48ad9
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/25057
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-boot-test: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Scott Murray <scott.murray@konsulko.com>