meta-agl-profile-graphical-html5: Point Chromium and WAM to agl-compositor
Updates the branches and revs to the Chromium and WAM repositories
with the ongoing changes to work with the new compositor.
It also depends on the integration of the changes on the
webapps to remove former dependencies, and an additional
fix to properly resume not visible webapps.

Bug-AGL: SPEC-3497

Signed-off-by: Lorenzo Tilve <ltilve@igalia.com>
Change-Id: Iddf1d3675810d718c670a6275b3d1d683461f252
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/25078
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-boot-test: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>