qtwayland-config: Do not enable by default resize after swap env variable
In our use-cases applications do not specify a size for the windows, leaving
qtwayland itself to create an initial size for the surface (a 500 by 500 size).
The compositor will send a configure event back to the application with the
available/actual geometry area, resulting in resize performed by the
application.

This, in tandem with the fact that we tweaked qtwayland to perform the resize
(due to SPEC-3334) after a buffer swap results in a short flicker -- the
application does an initial commit with that initial size but with no
background or any information attached to it. The environmental variable tunes
the repainting mechanism in qtwayland affecting the way the resize is performed
and was added when multiple or split surfaces are in use. Destroying one of
them will not cause a scale back to the original area.

We therefore disable by default that environment variable and leave it to the
applications to enable it at start-up, if they still want to use split type of
surfaces. Added a rather big comment about why that might necessary giving some
clues and some mention that qtwayland doesn't make a distinction for the value
of the environmental variable itself.

Bug-AGL: SPEC-3522

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Ib0dcda6382dd52677a0768afe3948ab62955c9f8
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/25120
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>