waltham_%.bbappend: Added bbappend recipe for waltham
Waltham no longer compiles and it is required for
waltham-transmitter-plugin to function. A follow-up task has been added
at SPEC-3647 to further determine what we should do about it. For now,
this adds a bbappend recipe and with it 2 patches that fixes trivial
compilation errors. The third patch is for 'enhancing' the waltham
protocol to pass from the transmitter plug-in to the receiver side
the app_id.

Bug-AGL: SPEC-3601, SPEC-3647

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Ie01e4ad1db0a64d9d911308250ea63a79f78c6fa
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/25452
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>