Add AGL reference hardware support to h3ulcb/h3ulcb-nogfx
Update the h3ulcb and h3ulcb-nogfx machine templates to pull in the
meta-agl-refhw layer and a new agl_refhw-h3.inc file that enables
the reference hardware support.  The result is an extra devicetree
file that can be used to boot the h3ulcb images on the reference
hardware.

Bug-AGL: SPEC-3658

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I5bbfd3a6014affb1c9668d1bce9378d84c4c76cb