virtualization/linux-yocto: Add virtio-aarch64-standard.scc to SRC_URI.
It allows to use patch command to apply patches via scc file.

As explained in [1,2], Yocto kernel metadata is commonly used for two
purposes:

1. To construct Yocto kernel git by committing patches listed in scc
files.
2. To configure Yocto kernel during Yocto build.

When scc file is not explicitly added to SRC_URI, then patch commands
will be ignored during Yocto build because it is assumed that patches
are already pre-applied during Yocto kernel git tree construction.

On the other hand, nopatch should be added to standard.scc because
patches from it are actually pre-applied.

[1]: https://www.yoctoproject.org/pipermail/yocto/2019-October/046986.html
[2]: https://docs.yoctoproject.org/kernel-dev/maint-appx.html

Change-Id: Id238e79e1b354daaa23e83bef45fc93a216fbf92
Bug-AGL: SPEC-3893
Signed-off-by: Vasyl Vavrychuk <vasyl.vavrychuk@opensynergy.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/26288
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>