meta-agl-bsp: Add reference hardware GPS support
Add recipe for gpsd configuration for the reference hardware, which
also installs a script plus systemd unit drop-in to handle enabling
and disabling it.  The latter is required due to the reference
hardware's reuse of the h3ulcb image, which means the existing
PREFERRED_PROVIDER_virtual/gpsd-conf scheme is not sufficient in
this case.

Bug-AGL: SPEC-3792

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ifaf3b93d06825f7c5c6a8ce2773d93931e47178c
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/25994
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>