agl-compositor: Bump SRCREV
This update is needed to allow unregistered/unknown applications in/to the
RBAModel json to be displayed. By default no application would be allowed. This
can be tweaked by modifying the weston.ini file. By default it is off/not set.

In detail this update includes the following the changes:

- fd860492da6fb07f Allow unregistered applications in rba
- 47386558dbc95387 Use parameterized constructor of RBAArbitrator
- 79bc7999332d55b7 Update architecture diagram with a small legend

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I6249fd934868b6c064cf9f183e3db5282ffdc2bf
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/26024
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>