recipes-graphics/wayland/agl-compositor: SRCREV bump
The major change in this bump, besides some smaller fixes, is that the
compositor would send to the client (before attaching a buffer to a surface)
the initial surface dimensions, thus not needing any resizing code for the
client to handle, making this a nice optimization to have to in.
Obviously, this happens only for regular clients, w/ the client shell,
still required to handle the resizing of its buffers.

With this bump we also remove an external dependency patch (which should
come as a follow-up to this) and increase the compositor version.

In detail the following changes have been added:

- 68d184f9d541a47 meson.build: Do a release update before libweston9 integration
- 483d363b9e724a0 layout: Remove weston_output_damage dependecy
- e35c7858a788b46 layout: Fix split role sending invalid dimensions
- 80febb3fbcebe88 desktop: Tell clients to use current geometry area from start
- 46e2d5e441dfb9f desktop: Send remove event only if output and app_id is set
- d85915199c8a321 Notify about application destruction
- fd860492da6fb07 Allow unregistered applications in rba

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Iaed23aae27695db51260b7b04e3d9e0a5a053aa2
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/26278
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>