rcar3: move extra kernel modules to MACHINE_EXTRA_RRECOMMENDS
The appending of the rcar3 graphics related kernel modules to
IMAGE_INSTALL in agl_rcar.inc is problematic for users building
headless or container targetted images.  To avoid this, instead
append them to MACHINE_EXTRA_RRECOMMENDS, which will result in
them being pulled in via packagegroup-machine-base.  Users not
desiring them can then either disable package recommendations
or not pull in packagegroup-machine-base.  In practice, this
has no effect on any of the current images in meta-agl-core or
meta-agl-demo, but it is required for the IC EG guest container
build for rcar3 platforms.

Bug-AGL: SPEC-3983

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I309926b6dc31614fb7c710b0a302716942d07df8
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/26430
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>