Remove default inclusion of kernel-modules
Remove the kernel-modules meta package append to IMAGE_INSTALL
from the default base-agl.inc aglsetup.sh template, as it essentially
turns a developer convenience into default behavior that may not be
desirable by all users.  Note that this does change the contents of
agl-image-(boot|minimal|weston) with respect to kernel modules on
most platforms, but if this proves problematic it is recommended that
kernel-modules be added back elsewhere in a more transparent fashion.
A corresponding change will be made in meta-agl-demo to retain the
behavior for the demo platform images.

Bug-AGL: SPEC-3983

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ie1098f25d69e721345d13db3681278543d594e78
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/26431
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>