agl-compositor_git: Bump SRCREV
This bump brings no functional change, it just re-factors some of the
code in the compositor to match a bit more to what upstreaming into has
in weston, and paves the way to add support for testing, and with it, to
include headless backend loading. The bump would need changes in the
recipe so it will have a couple of follows-up, to compile the headless
back-end and install a run-time shared library.

Note this change squashed headless-backend support as well to make CI
happy (Bug-AGL: SPEC-3964) and Install the run-time shared library (Bug-AGL: SPEC-3964)

In more detail, the following changes have been included:

- de7cdb0314af949 compositor: Avoid loading systemd-notify plug-in
- ae60166cd931609 meson.build: Install run-time shared library to a specific path
- 0bef8700f21b9e7 main: Split of main() from wet_main()
- 65361f9f7979056 main: Copy cmd line args and print them
- ccda88821b5bd9a main: Teach agl-compositor to load additional modules
- 61dff3ff317c4d6 main: Add support for loading the headless-backend

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I49bbb216cb723db6a104d9e2753f5ddeb9bdd9f0
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/26442
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>