Update license handling for pipewire recipe
We need to set LICENSE to multiple (MIT & GPL-2.0-only & LGPL-2.1-or-later) to satisfy
the current scheme required by bitbake.

But the licenses of the output packages are not all 3,
but mostly MIT and a few LGPL-2.1-or-later and one GPL-2.0-only.
To map the dynamically created output packages with the proper MIT license,
we need to call setVar after all dynamic packages are created.

Bug-AGL: SPEC-3844
Signed-off-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Change-Id: I0900902af2527254869c125ec90239c9afb26f3c
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/26449
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Georgios Kiagiadakis <george.kiagiadakis@collabora.com>