Update license handling for pipewire recipe
We need to set LICENSE to multiple (MIT & GPL-2.0-only & LGPL-2.1-or-later) to satisfy
the current scheme required by bitbake.

But the licenses of the output packages are not all 3,
but mostly MIT and a few LGPL-2.1-or-later and one GPL-2.0-only.
To map the dynamically created output packages with the proper MIT license,
we need to call setVar after all dynamic packages are created.

Bug-AGL: SPEC-3844

Signed-off-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
[manual cherry-pick of 340e5d4c from master/lamprey]
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ie27a929a35e713c7478cde8b097ea1da51428047
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/26571
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account