qt5: Fix timer leak in qtwayland to avoid animations being sluggish
With long-running qt applications which have fluid animations
in wayland environment, the animation becomes obviously sluggish
because massive amount of memcpy() is called through a constructor
being placed in a loop in QTimerInfoList::timerInsert() function.
This is caused by a timer-index leak bug of qtwayland, which is
already reported to the qt project as following ticket.

https://bugreports.qt.io/browse/QTBUG-79838

Unfortunately QTBUG-79838 is still open because maintenance of
non-commercial version of qt5 is already stopped. However, we've
confirmed that the patch attached to that ticket works fine
except for that part of it is no longer needed due to the deletion
of the code. So let's apply only effective part of it.

Bug-AGL: SPEC-3991

Signed-off-by: Hiroyuki Ishii <ishii.hiroyuki002@jp.panasonic.com>
Change-Id: Ib148b81aabb98e8df10c1414cdbfe26f7ddf09a6
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/26466
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>