meta-agl: split wireplumber to run in multiple instances
The "IC SoundManager" design specifies that there needs to be
a session manager running in the host to manage devices and permissions,
but all the domain-specific policy and app management needs to run in
the IVI container, together with the apps. With this split, we are able
to achieve that.

In addition to the "host" and "policy" instances, we are also adding
a "bluetooth" one, which loads the bluetooth plugin and manages the
bluetooth device. This can be moved to the IVI container as well,
or elsewhere... it only depends on bluez (so it must run in the same
container as bluez).

For now, given the absence of an IVI container in the lxc-demo image,
all instances are running in the host, but it is trivial to move all
the non-host ones to another container later.

To compliment pipewire-ic-ipc, this change also adds an "alsa-suspend"
lua script that runs in the context of the "host" wireplumber instance
and its purpose is to mute pipewire-managed alsa devices when there is
a sound playing in the IC container (on another alsa device).

Finally, this change also adds V4L2 configuration in the "host" wireplumber
instance, which is still disabled (and untested), but you can easily enable
it for experimentation by uncommenting the relevant line in
host.lua.d/90-enable-all.lua

Bug-AGL: SPEC-4027
Signed-off-by: George Kiagiadakis <george.kiagiadakis@collabora.com>
Change-Id: I9febc4f3919e7c559a5d7d32bfe7bc95c75934f2
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/26662
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Naoto YAMAGUCHI <naoto.yamaguchi@aisin.co.jp>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>