meta-app-framework: add new AGL feature and fix build
This layer uses a new `agl-app-fw` feature to guard bbappends, so this
commit declares it in the conf include and makes sure it's used when the
feature is enabled.

Moreover, the `layer.conf` files referenced a non-existent `licenses`
folder, causing build issues. This commit fixes this by not adding this
folder to LICENSE_PATH.

Bug-AGL: SPEC-4161
Change-Id: I3360b0f3b36413e741de2d2d3e4a98fff1e1e8db
Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/26973
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account