meta-app-framework: start agl-compositor in a systemd user session
`weston-init` installs a system service for starting the compositor. As
we want it to run as a user service, this patch replaces this package
with `agl-compositor-init`. This involves:
- resetting RDEPENDS from `weston-ini-conf` so it doesn't pull in
  `weston-init` anymore
- making `agl-compositor-init` RCONFLICT with `weston-init` as it's
  still a RRECOMMENDS to `weston`

`agl-compositor-init` installs a systemd user service and socket, and
ensures agl-compositor is automatically started with the user session,
so the compositor is automatically started on boot. It also adds a
config fragment for the `agl-session@agl-driver` service (instance of
`agl-session` for user `agl-driver`) for properly configuring the TTY to
use for display.

Bug-AGL: SPEC-4161
Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>
Change-Id: Ifb10a206cfa97c147283bc78c46f7f09209a9591
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/26975
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
Reviewed-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
ci-image-boot-test: Jenkins Job builder account