agl-compositor-init_agl-app-fw: Include agl-shell-activator
Turns out we weren't including the agl-shell-activator activation
tool so include it by default. It remains to be seen if it makes
sense to have in the compositor, rather than outside, but for the time
being  just include it by default.

Bug-AGL: SPEC-4324
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I5a16baca8be83715f86e02809324497a837a235a
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/27385
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Vasyl Vavrychuk <vasyl.vavrychuk@opensynergy.com>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>