meta-agl-core: switch next branch users back to master
Remove the aglnext override that was added on the next branch to
switch the couple of affected libweston users (agl-compositor and
waltham-transmitter-plugin) to forked next branches for handling
Weston 9.0 and 10.0 changes.  Now that those changes have been
merged to the master branches in the respective repositories, this
is no longer required.  As well, bump the SRCREVs to pick up the
Weston update changes.

Bug-AGL: SPEC-4343

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I7a4598a9e776aefa3ec7b2bf13b2e89c086c0681
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/27418