meta-agl-bsp: avoid meta-gnome dependency from meta-rcar-gen3
The kirkstone-dev_3.5_M3_v5.9.0 branch of meta-renesas contains a
new bbappend for cogl which requires meta-gnome.  Since nothing
in the BSP layer itself depends on cogl, and there is currently
no obvious usecase for cogl outside of GNOME desktop related
libraries, BBMASK out the bbappend so we do not need to pull in
meta-gnome.  This can be revisited if an AGL member presents a need
for cogl on rcar3 in their downstream use of AGL.  Ideally, upstream
will move the bbappend to a dynamic layer directory to remove the
issue.

Bug-AGL: SPEC-4343

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I6e55c07d1ff1fd7d68d1798ecca3c9c1b35cdf14
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/27428
Reviewed-by: Vasyl Vavrychuk <vasyl.vavrychuk@opensynergy.com>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>