meta-agl-bsp: fix rcar3 packaging errors
Provide a default value of PACKAGE_ARCH for the packagegroup
packagegroup-multimedia-libs from meta-rcar-gen3 to avoids errors from
a new packaging QA check with respect to packagegroups depending on
versioned library packages needing to be machine-specific.

Note that this needs to be done in global configuration space as
opposed to in a bbappend because the value needs to be set before the
packagegroup bbclass is inherited.  This can be removed when upstream
is fixed to add the required PACKAGE_ARCH = "${MACHINE_ARCH}".

Bug-AGL: SPEC-4343

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I60dc1fbb976c7e08b174017268949dc43d6d17fe
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/27429
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>