meta-agl-bsp: fix building weston for rcar3 nogfx machines
Work around use of obsolete Weston backend-fbdev in meta-rcar-gen3
when building without the proprietary driver stack.  Masking out the
weston_%.bbappend seems the simplest approach for now, as forcing
USE_GLES = "1" to disable the adding of -Dbackend-default="fbdev" to
EXTRA_OEMESON triggers adding "libgbm" to DEPENDS, which breaks due
to that being a component specific to the proprietary stack we are
explicitly disabling.  If the "libgbm" usage in the bbappend was
instead "virtual/libgbm" as is now used in poky, it might be possible
to just tweak PREFERRED_PROVIDER_virtual/libgbm = "mesa".

Bug-AGL: SPEC-4343

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I603a9c4744befb97513808714cedd3e21664525f
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/27430
ci-image-build: Jenkins Job builder account
Reviewed-by: Vasyl Vavrychuk <vasyl.vavrychuk@opensynergy.com>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>