Add template for generic-arm64 machine
This is a standard aarch64 (ARM 64bit) machine for use with
'SystemReady' (aka UEFI) systems.  It had been added to the next
branch in meta-agl-devel for some testing on AWS Graviton instances,
but now that the upgrade to kirkstone has landed, it seems sensible
to move it to meta-agl to ease maintenance.

Bug-AGL: SPEC-4353

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I51e223ee80ea62bce99b725c4daef1a010cb0835
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/27456
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account