meta-agl-core: move polkit addition to dynamic-layers
Since polkit comes from meta-oe and not poky, move adding it to
packagegroup-agl-graphical-weston to a bbappend that is in a meta-oe
dynamic-layers hierarchy.  This is required to avoid breaking the
meta-agl-core testing on the Yocto autobuilder, which does not pull
in meta-oe currently.  This may need to be revisited down the road
when providing a binary packagefeed is attempted, as it may be the
case that explicitly adding meta-oe as a meta-agl-core dependency
ends up the better solution for packagefeed stability.

Bug-AGL: SPEC-4359, SPEC-4375

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I920d9184f936d0c529db944bf1667a470aa863c6
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/27568
Reviewed-by: Marius Vlad <marius.vlad@collabora.com>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account