weston-ini-conf: rework remote output configuration generation
To allow for a user to change the weston-remoting configuration
fragment the same way 8c4b3948 did with the the fragment for
waltham-remoting, rename the TRANSMITTER* variables 8c4b3948
added to REMOTE* and apply the same transformations to the
remote-output.cfg fragment.  Additionally, add a new
REMOTE_OUTPUT_APP_ID variable that will add the potentially
required agl-shell-app-id option to the fragment if the variable
is set.

Bug-AGL: SPEC-4426

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ib20aafe5737482814fe6c4ae0f9cd1b39230adb3
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/27619
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account