meta-agl-core: add connman-conf dependency
After kirkstone, connman has an independent package for its
configuration. This package is not added as a dependency.

As a consequence, we do not ship /etc/connman/main.conf, that, in
AGL builds, is expected to provide a fallback for NTP. This means
Raspberry PI4 does not synchronize the clock on boot if connection
is available. As a side effect, all SSL connections fail.

Bug-AGL: SPEC-4447
Change-Id: Iaa34c0a41cd161b0a0edea7602c39f724831412a
Signed-off-by: Jose Dapena Paz <jdapena@igalia.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/27683
Reviewed-by: Marius Vlad <marius.vlad@collabora.com>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account