x86-extra-graphic-devices: Build vmwgfx driver as built-in
Rather than building the vmwgfx driver as a module, build it as
a built-in in module to avoid a possible probe failure when the driver
loads up.

This is a temporary work-around which seems to work just fine on VBox
but also on qemu.

Bug-AGL: SPEC-4449
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I4f342854df1a0ed2d14ed0ca2bbe754618c61689
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/27695
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>