Fix a dependency issue when building the SDK
When using -c populate_sdk, then the -dev packages are installed.
There is a hard dependency of the -dev package on the  package.
As we have multiple subtypes here and switch using ALTERNATIVES, this
collides in this corner case.
Fix it by not creating a -dev package in the first place as
it is not needed here anyway.

Signed-off-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Change-Id: I21326c9edc5cb7571ed94593eae1075cf56353db
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/27841
Reviewed-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-by: Marius Vlad <marius.vlad@collabora.com>
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account