weston_10.0_aglcore.inc: Update wl_output to version 4
This allows clients to query the output name and description name as
seen by the compositor. Previously this was only possible only with
xdg_output_manager extension but a recent update to the core wayland
protocol allows us to do just that.

Patch is posted upstream
https://gitlab.freedesktop.org/wayland/weston/-/merge_requests/987, but
will take some time to get in, as we're in the process of a release.

Bug-AGL: SPEC-4530
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I709b941a220e9906078944ee585435df7141d448
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/27949
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>