agl-compositor_git.bb: Bump SRCREV
This SRCREV bump update brings in a helpful way to designate a certain
area such that the compositor would use it when activating applications,
rather than using the entire output. When not having panels the
background surface would cover the entire output, making it
impossible to keep visible areas of the background surface and, in
the same time, display another's application surface.

While the solution would be use panels specifically designed to handle
this case, that's not always possible.  We circumvent that by having this
area specified manually in this update, with a direct follow-up to have
a request in the agl-shell protocol that specifies that area
programmatically.

Besides that change, a few more changes made it through, all of them being
added from upstream and adapted from weston to the AGL compositor,
mostly related to how we handle signals.

In detail the following changes made it through since the last SRCREV
bump update:

- 924473ef016b Add manual activation area configuration option
- 897449688a9c compositor: Use sigaction to trap SIGINT
- e8761762fcd0 compositor: Add missing SIGCHLD handler
- 315496ddbb9f compositor: Stop trapping SIGQUIT
- f980c0d90329 shell: Rewrite client_exec as client_launch
- edb6d397a4c6 process-util: Move Xwayland fork helpers to shared
- 86146a4e7455 shared: add str_printf()
- 1b93121761e1 shared: introduce os_fd_clear_cloexec()

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: I23e0aba9f2b77cab4fe81754a8feaeebb7cea340
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/28056
ci-image-build: Jenkins Job builder account
Tested-by: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>